sample/requirements.txt
twelvedata
websocket-client
psycopg2-binary
python-dotenv
sqlalchemy